Shift logic back to Gui class
