+++ src/main/java/io/gui/MainWindow.java
@@ -10,7 +10,6 @@ import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
 import task.TaskList;
@@ -30,19 +29,18 @@ public class MainWindow extends AnchorPane implements Ui {
     @FXML
     private Button sendButton;
 
-    private TaskList taskList;
-    private Storage<TaskList> storage;
+    Gui gui;
 
     private final Image userImg = new Image(this.getClass().getResourceAsStream("/images/default_user_pic.png"));
     private final Image dImg = new Image(this.getClass().getResourceAsStream("/images/cyborg_girl.png"));
 
+    /**
+     * Setup of tasks and storage.
+     */
     @FXML
     public void initialize() {
         scrollPane.vvalueProperty().bind(dialogueContainer.heightProperty());
         showReply(WELCOME);
-        storage = Storage.of(TaskList.class, "taskList.ser");
-        loadTasks();
-        showReply(String.format("Current tasks: %s", taskList.toString()));
     }
 
     /**
@@ -62,7 +60,7 @@ public class MainWindow extends AnchorPane implements Ui {
         Command command = Command.parser().parse(userText).match(
                 pr -> pr.first(),
                 msg -> Error.of(msg));
-        command.execute(taskList, this, storage);
+        command.execute(gui.taskList, this, gui.storage);
         if (command.isExit()) {
             Platform.exit();
         }
@@ -74,23 +72,5 @@ public class MainWindow extends AnchorPane implements Ui {
     @Override
     public void showReply(String msg) {
         dialogueContainer.getChildren().addAll(DialogueBox.of(msg, dImg, dialogueContainer));
-    }
-
-    private void loadTasks() {
-        storage = Storage.of(TaskList.class, "taskList.ser");
-        taskList = storage.load().match(
-                lst -> lst,
-                error -> {
-                    switch (error) {
-                        case FILE_NOT_FOUND:
-                            return new TaskList();
-                        case IO_ERROR:
-                        case CAST_ERROR:
-                            showReply("Error loading tasks, tasks have been reset");
-                            return new TaskList();
-                        default:
-                            return new TaskList();
-                    }
-                });
     }
 }
